Have an unsafe __allowNetworking flag
When Nix is configured with `allow-allow-networking = true` and the
derivation has `__allowNetworking = true` set, network access is
allowed during the build.

This is mainly useful when using Nix as a deployment system. For
example this allows derivations that provision resources using
terraform like:

```
with (import <nixpkgs> {});
stdenv.mkDerivation {
  name = "deploy-resource";
  src = ./.;
  phases = ["unpackPhase" "installPhase"];
  impureEnvVars = [ "AWS_ACCESS_KEY_ID" "AWS_SECRET_ACCESS_KEY" ];
  nativeBuildInputs = [ terraform ];
  __allowNetworking = true;
  installPhase = ''
    terraform init
    terraform apply -auto-approve
    touch $out
  '';
}
```

Although the derivation is not fully deterministic (because terraform
will query some remote state) it's still very hard to miss specifying
a required dependency.